test(server): split test files
